[development-minor-8]
1. minor corrections to the fitter functions
2. docstrings added to `powerLaw_d.py` and `powerLaw_f.py`

+++ src/pymultifit/fitters/gaussian_f.py
@@ -12,22 +12,22 @@ from ..distributions.gaussian_d import gaussianWA
 
 class GaussianFitter(BaseFitter):
     """A class for fitting multiple Gaussian functions to the given data."""
-
+    
     def __init__(self, n_fits: int, x_values, y_values, max_iterations: int = 1000):
         super().__init__(n_fits=n_fits, x_values=x_values, y_values=y_values, max_iterations=max_iterations)
         self.n_par = 3
-
+    
     @staticmethod
     def _fitter(x: np.ndarray, params: list) -> np.array:
         return gaussianWA(*params).pdf(x)
-
+    
     def _n_fitter(self, x: np.ndarray, *params: Any) -> np.ndarray:
-        y = np.zeros_like(x)
+        y = np.zeros_like(x, dtype=float)
         params = np.reshape(params, (self.n_fits, self.n_par))
         for amp, mu, sigma in params:
             y += self._fitter(x=x, params=[amp, mu, sigma])
         return y
-
+    
     def _plot_individual_fitter(self, x: np.ndarray, plotter):
         params = np.reshape(self.params, (self.n_fits, self.n_par))
         for i, (amp, mu, sigma) in enumerate(params):
@@ -36,17 +36,17 @@ class GaussianFitter(BaseFitter):
                                      f'{self.format_param(amp)}, '
                                      f'{self.format_param(mu)}, '
                                      f'{self.format_param(sigma)})')
-
+    
     @staticmethod
     def format_param(value, threshold1=0.001, threshold2=10_000):
         """Formats the parameter value based on its magnitude."""
         return f'{value:.3E}' if threshold2 < abs(value) or abs(value) < threshold1 else f'{value:.3f}'
-
+    
     def _get_overall_parameter_values(self) -> tuple[list, list]:
         _, mu, _ = self.parameter_extractor(mean=True)
         amp = get_y_values_at_closest_x(x_array=self.x_values, y_array=self.get_fit_values(), target_x_values=mu)
         return amp, mu
-
+    
     @deprecated("This function will be dropped in upcoming releases. Use get_parameters function instead")
     def parameter_extractor(self,
                             amplitude: bool = False, mean: bool = False, standard_deviation: bool = False,
@@ -77,15 +77,15 @@ class GaussianFitter(BaseFitter):
         """
         if not (amplitude or mean or standard_deviation):
             return [], [], []
-
+        
         values = self.get_value_error_pair(mean_values=True)
-
+        
         if fit_amplitude:
             amp_values, mu_values = self._get_overall_parameter_values()
             return amp_values, mu_values, []
-
+        
         amp_values, mu_values, sigma_values = [], [], []
-
+        
         n_fits, n_par = self.n_fits, self.n_par
         if amplitude:
             amp_values = [values[i * n_par] for i in range(n_fits)]
@@ -93,19 +93,19 @@ class GaussianFitter(BaseFitter):
             mu_values = [values[i * n_par + 1] for i in range(n_fits)]
         if standard_deviation:
             sigma_values = [values[i * n_par + 2] for i in range(n_fits)]
-
+        
         return amp_values, mu_values, sigma_values
-
+    
     def get_parameters(self,
                        amplitude: bool = False, mu: bool = False, sigma: bool = False,
                        gaussian_number: tuple = None):
         if not (amplitude or mu or sigma):
             return [], [], []
-
+        
         values = self.get_value_error_pair(mean_values=True)
-
+        
         amp_values, mu_values, sigma_values = [], [], []
-
+        
         n_fits, n_par = self.n_fits, self.n_par
         if amplitude:
             amp_values = [values[i * n_par] for i in range(n_fits)]
@@ -119,5 +119,5 @@ class GaussianFitter(BaseFitter):
             sigma_values = [values[i * n_par + 2] for i in range(n_fits)]
             if gaussian_number:
                 sigma_values = [sigma_values[i - 1] for i in gaussian_number]
-
+        
         return np.array(amp_values), np.array(mu_values), np.array(sigma_values)
