+++ src/pymultifit/fitters/logNorm_f.py
@@ -29,7 +29,7 @@ class LogNormalFitter(BaseFitter):
         return log_normal_(x, *params, normalize=False)
 
     def _n_fitter(self, x, *params):
-        y = np.zeros_like(x)
+        y = np.zeros_like(x, dtype=float)
         params = np.reshape(params, (self.n_fits, self.n_par))
         for amp, mu, sigma in params:
             y += self._fitter(x, [amp, mu, sigma])
