+++ src/pymultifit/fitters/skewNorm_f.py
@@ -26,7 +26,7 @@ class SkewedNormalFitter(BaseFitter):
         return params[0] * skewnorm.pdf(x, params[1], loc=params[2], scale=params[3])
 
     def _n_fitter(self, x, *params):
-        y = np.zeros_like(x)
+        y = np.zeros_like(x, dtype=float)
         params = np.reshape(params, (self.n_fits, self.n_par))
         for amp, shape, loc, scale in params:
             y += self._fitter(x, [amp, shape, loc, scale])
